+++ python/optim.py
@@ -251,11 +251,6 @@ class SGD(torch.optim.SGD):
         state["velocity_accum_type"] = self.velocity_accum_type
         return state
 
-    def state_dict(self) -> Dict[str, Any]:
-        logger.warning(
-            "IPU-specific optimizer states cannot be read from the host.")
-        return super().state_dict()
-
 
 class Adam(torch.optim.Adam):
     """ Adam optimizer.
@@ -351,11 +346,6 @@ class Adam(torch.optim.Adam):
         state["second_order_momentum_accum_type"] = \
                 self.second_order_momentum_accum_type
         return state
-
-    def state_dict(self) -> Dict[str, Any]:
-        logger.warning(
-            "IPU-specific optimizer states cannot be read from the host.")
-        return super().state_dict()
 
 
 class AdamW(torch.optim.AdamW):
@@ -457,11 +447,6 @@ class AdamW(torch.optim.AdamW):
         state["second_order_momentum_accum_type"] = \
                 self.second_order_momentum_accum_type
         return state
-
-    def state_dict(self) -> Dict[str, Any]:
-        logger.warning(
-            "IPU-specific optimizer states cannot be read from the host.")
-        return super().state_dict()
 
 
 class RMSprop(torch.optim.RMSprop):
@@ -573,11 +558,6 @@ class RMSprop(torch.optim.RMSprop):
                 self.second_order_momentum_accum_type
         state["use_tf_variant"] = self.use_tf_variant
         return state
-
-    def state_dict(self) -> Dict[str, Any]:
-        logger.warning(
-            "IPU-specific optimizer states cannot be read from the host.")
-        return super().state_dict()
 
 
 class LAMB(torch.optim.Optimizer):
@@ -745,11 +725,6 @@ class LAMB(torch.optim.Optimizer):
         state["second_order_momentum_accum_type"] = \
                 self.second_order_momentum_accum_type
         return state
-
-    def state_dict(self) -> Dict[str, Any]:
-        logger.warning(
-            "IPU-specific optimizer states cannot be read from the host.")
-        return super().state_dict()
 
 
 def _check_constructor_match_parent(child_class: Type[torch.optim.Optimizer]